Fix capitalization
